[bugfix] use after free & fix travis (#2268)
* fix use after free

* fix travis

* fix log

+++ native/tools/travis-scripts/before-install.sh
@@ -31,11 +31,16 @@ function install_android_ndk()
 
 function install_clang()
 {
-    echo "Download clang"
-    curl -O http://releases.llvm.org/5.0.0/clang+llvm-5.0.0-linux-x86_64-ubuntu14.04.tar.xz
-    echo "Decompress clang"
-    tar xpf ./clang+llvm-5.0.0-linux-x86_64-ubuntu14.04.tar.xz
-    cp ./clang+llvm-5.0.0-linux-x86_64-ubuntu14.04/lib/libclang.so.5.0 $COCOS2DX_ROOT/tools/bindings-generator/libclang/libclang.so
+    if [ ! -f $COCOS2DX_ROOT/tools/bindings-generator/libclang/libclang.so ]; then
+        echo "Download clang"
+        curl -O http://releases.llvm.org/5.0.0/clang+llvm-5.0.0-linux-x86_64-ubuntu14.04.tar.xz
+        echo "Decompress clang"
+        tar xpf ./clang+llvm-5.0.0-linux-x86_64-ubuntu14.04.tar.xz
+        cp ./clang+llvm-5.0.0-linux-x86_64-ubuntu14.04/lib/libclang.so.5.0 $COCOS2DX_ROOT/tools/bindings-generator/libclang/libclang.so
+    else
+        echo "Skip downloading clang"
+        echo "  file $COCOS2DX_ROOT/tools/bindings-generator/libclang/libclang.so exists!"
+    fi
 }
 
 function install_python_module_for_osx()
